Updated routes and templates

+++ application/admin_routes.py
@@ -4,12 +4,16 @@ from .models  import User,Booking,Staff,Trek
 from .db import db
 from datetime import datetime
 
+
+
 @app.route("/admin/dashboard")
 def admin():
+    if("admin_id" not in session):
+        return redirect("/login")
     this_admin=User.query.filter_by(id=session["admin_id"]).first()
-    total_user=User.query.all().__len__()
     total_treks=Trek.query.all().__len__()
     total_staff=Staff.query.all().__len__()
+    total_user=User.query.all().__len__()-total_staff-1
     all_bookings=Booking.query.all()
     recent_booking=Booking.query.order_by(Booking.booking_date.desc()).limit(3).all()
     total_bookings=all_bookings.__len__()
@@ -22,12 +26,16 @@ def admin():
 
 @app.route("/admin/treks")
 def admin_trek():
+    if("admin_id" not in session):
+        return redirect("/login")
     this_admin=User.query.filter_by(id=session["admin_id"]).first()
     all_treks=Trek.query.all()
     return render_template("admin/admin_treks.html",this_admin=this_admin,all_treks=all_treks)
 
 @app.route("/trek/add",methods=["GET","POST"])
 def add_trek():
+    if("admin_id" not in session):
+        return redirect("/login")
     this_admin=User.query.filter_by(id=session["admin_id"]).first()
     ex_staff=Staff.query.all()
     if(request.method=="POST"):
@@ -42,7 +50,7 @@ def add_trek():
         staff=request.form.get("staff")
         status=request.form.get("status")
         desc=request.form.get("desc")
-        trek=Trek(name=name,location=location,price=price,difficulty=difficulty,duration_days=duration_days,total_slots=total_slots,avl_slots=total_slots,start_date=start_date,end_date=end_date,assigned_staff_id=staff,created_by=this_admin.id,desc=desc)
+        trek=Trek(name=name,location=location,price=price,difficulty=difficulty,duration_days=duration_days,total_slots=total_slots,avl_slots=total_slots,start_date=start_date,end_date=end_date,assigned_staff_id=staff,created_by=this_admin.id,desc=desc,status=status)
         db.session.add(trek)
         db.session.commit()
     return render_template("admin/add_trek.html",this_admin=this_admin,ex_staff=ex_staff)
@@ -50,6 +58,8 @@ def add_trek():
 
 @app.route("/edit/trek/<int:trek_id>",methods=["GET","POST"])
 def edit_trek(trek_id):
+    if("admin_id" not in session):
+        return redirect("/login")
     this_trek=Trek.query.filter_by(id=trek_id).first()
     this_admin=User.query.filter_by(id=session["admin_id"]).first()
     ex_staff=Staff.query.join(User,Staff.userid==User.id).all()
@@ -76,6 +86,8 @@ def edit_trek(trek_id):
 
 @app.route("/delete/trek/<int:trek_id>")
 def delete_trek(trek_id):
+    if("admin_id" not in session):
+        return redirect("/login")
     this_trek=Trek.query.filter_by(id=trek_id).first()
     db.session.delete(this_trek)
     db.session.commit()
@@ -84,6 +96,8 @@ def delete_trek(trek_id):
 
 @app.route("/admin/staff")
 def all_staff():
+    if("admin_id" not in session):
+        return redirect("/login")
     this_admin=User.query.filter_by(id=session["admin_id"]).first()
     all_staff=Staff.query.all()
     active_staff=Staff.query.filter_by(approval_status="approved").all().__len__()
@@ -95,6 +109,8 @@ def all_staff():
 
 @app.route("/approve_staff/<int:staff_id>")
 def approve_staff(staff_id):
+    if("admin_id" not in session):
+        return redirect("/login")
     this_admin=User.query.filter_by(id=session["admin_id"]).first()
     this_staff=Staff.query.filter_by(id=staff_id).first()
     this_staff.approval_status="approved"
@@ -105,6 +121,8 @@ def approve_staff(staff_id):
 
 @app.route("/reject_staff/<int:staff_id>")
 def reject_staff(staff_id):
+    if("admin_id" not in session):
+        return redirect("/login")
     this_admin=User.query.filter_by(id=session["admin_id"]).first()
     this_staff=Staff.query.filter_by(id=staff_id).first()
     this_staff.approval_status="blacklisted"
@@ -114,6 +132,8 @@ def reject_staff(staff_id):
 
 @app.route("/admin/all/booking")
 def all_booking():
+    if("admin_id" not in session):
+        return redirect("/login")
     this_admin=User.query.filter_by(id=session["admin_id"]).first()
     all_bookings=Booking.query.all()
     return render_template("admin/all_booking.html",all_bookings=all_bookings,this_admin=this_admin)
@@ -122,12 +142,16 @@ def all_booking():
 
 @app.route("/admin/users")
 def all_users():
+    if("admin_id" not in session):
+        return redirect("/login")
     this_admin=User.query.filter_by(id=session["admin_id"]).first()
     all_users=User.query.filter_by(role="user").all()
     return render_template("admin/all_users.html",all_users=all_users,this_admin=this_admin)
 
 @app.route("/revoke/<int:user_id>")
 def revoke_user(user_id):
+    if("admin_id" not in session):
+        return redirect("/login")
     this_admin=User.query.filter_by(id=session["admin_id"]).first()
     this_user=User.query.filter_by(id=user_id).first()
     this_user.status="blacklisted"
@@ -137,6 +161,8 @@ def revoke_user(user_id):
 
 @app.route("/activate/<int:user_id>")
 def acticate_user(user_id):
+    if("admin_id" not in session):
+        return redirect("/login")
     this_admin=User.query.filter_by(id=session["admin_id"]).first()
     this_user=User.query.filter_by(id=user_id).first()
     this_user.status="active"
@@ -146,7 +172,8 @@ def acticate_user(user_id):
 
 @app.route("/admin/search")
 def admin_search():
-
+    if("admin_id" not in session):
+        return redirect("/login")
     search_type = request.args.get("type", "trek")
     query = request.args.get("q", "").strip()
     this_admin=User.query.filter_by(id=session["admin_id"]).first()
@@ -161,7 +188,8 @@ def admin_search():
 
         elif search_type == "user":
             results = User.query.filter(
-                User.name.ilike(f"%{query}%")
+                User.name.ilike(f"%{query}%"),
+                User.role=="user"
             ).all()
 
         elif search_type == "staff":
@@ -180,7 +208,19 @@ def admin_search():
 
 @app.route("/history/<int:user_id>")
 def user_history(user_id):
+    if("admin_id" not in session):
+        return redirect("/login")
+    this_admin=User.query.filter_by(id=session["admin_id"]).first()
     user=User.query.filter_by(id=user_id).first()
     history=Booking.query.filter_by(user_id=user_id).all()
-    this_admin=User.query.filter_by(id=session["admin_id"]).first()
+    
     return render_template("admin/history.html",user=user,history=history,this_admin=this_admin)
+
+
+@app.route("/admin/trek_details/<int:trek_id>")
+def trek_det(trek_id):
+    if("admin_id" not in session):
+        return redirect("/login")
+    this_admin=User.query.filter_by(id=session["admin_id"]).first()
+    trek=Trek.query.filter_by(id=trek_id).first()
+    return render_template("admin/trek_details.html",this_admin=this_admin,trek=trek)
